slave uses input mapping for function code 01 and 03
Changed mapping for function codes 01 and 03 to register/coil inputs.
The slave now uses the same mapping for read holding/read inputs.
Allows for one mapping to be used for typical slave applications with
function code 3/16 and different adresses for read/write. If read/write data
muste be the same, this can be solved in the user application.